fix(langchain): function calling, other parameters
