+++ packages/opentelemetry-instrumentation-langchain/tests/cassettes/test_llms/test_bedrock.yaml
