+++ packages/opentelemetry-instrumentation-langchain/tests/test_lcel.py
@@ -1,17 +1,10 @@
 import json
-import re
 
-import boto3
 import pytest
 from langchain.output_parsers.openai_functions import JsonOutputFunctionsParser
 from langchain.prompts import ChatPromptTemplate
 from langchain.prompts import PromptTemplate
 from langchain.schema import StrOutputParser
-from langchain_anthropic import ChatAnthropic
-from langchain_community.chat_models import BedrockChat
-from langchain_community.llms.huggingface_text_gen_inference import (
-    HuggingFaceTextGenInference,
-)
 from langchain_community.utils.openai_functions import (
     convert_pydantic_to_openai_function,
 )
@@ -232,190 +225,3 @@ async def test_async_invoke(exporter):
         "StrOutputParser.task",
         "RunnableSequence.workflow",
     ] == [span.name for span in spans]
-
-
-@pytest.mark.vcr
-def test_custom_llm(exporter):
-    prompt = ChatPromptTemplate.from_messages(
-        [("system", "You are a helpful assistant"), ("user", "{input}")]
-    )
-    model = HuggingFaceTextGenInference(
-        inference_server_url="https://w8qtunpthvh1r7a0.us-east-1.aws.endpoints.huggingface.cloud"
-    )
-
-    chain = prompt | model
-    response = chain.invoke({"input": "tell me a short joke"})
-
-    spans = exporter.get_finished_spans()
-
-    assert [
-        "ChatPromptTemplate.task",
-        "HuggingFaceTextGenInference.completion",
-        "RunnableSequence.workflow",
-    ] == [span.name for span in spans]
-
-    hugging_face_span = next(
-        span for span in spans if span.name == "HuggingFaceTextGenInference.completion"
-    )
-
-    assert hugging_face_span.attributes[SpanAttributes.LLM_REQUEST_TYPE] == "completion"
-    assert hugging_face_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "unknown"
-    assert (
-        hugging_face_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.content"]
-        == "System: You are a helpful assistant\nHuman: tell me a short joke"
-    )
-    assert (
-        hugging_face_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"]
-        == response
-    )
-
-
-@pytest.mark.vcr
-def test_openai(exporter):
-    prompt = ChatPromptTemplate.from_messages(
-        [("system", "You are a helpful assistant"), ("human", "{input}")]
-    )
-    model = ChatOpenAI(model="gpt-3.5-turbo")
-    chain = prompt | model
-
-    response = chain.invoke({"input": "Tell me a joke about OpenTelemetry"})
-
-    spans = exporter.get_finished_spans()
-
-    assert [
-        "ChatPromptTemplate.task",
-        "ChatOpenAI.chat",
-        "RunnableSequence.workflow",
-    ] == [span.name for span in spans]
-
-    openai_span = next(span for span in spans if span.name == "ChatOpenAI.chat")
-
-    assert openai_span.attributes[SpanAttributes.LLM_REQUEST_TYPE] == "chat"
-    assert openai_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "gpt-3.5-turbo"
-    assert (
-        openai_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.content"]
-    ) == "You are a helpful assistant"
-    assert (openai_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.role"]) == "system"
-    assert (
-        openai_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.1.content"]
-    ) == "Tell me a joke about OpenTelemetry"
-    assert (openai_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.1.role"]) == "user"
-    assert (
-        openai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"]
-        == response.content
-    )
-    assert (
-        openai_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.role"]
-    ) == "assistant"
-
-
-@pytest.mark.vcr
-def test_anthropic(exporter):
-    prompt = ChatPromptTemplate.from_messages(
-        [("system", "You are a helpful assistant"), ("user", "{input}")]
-    )
-    model = ChatAnthropic(model="claude-2.1")
-
-    chain = prompt | model
-    response = chain.invoke({"input": "tell me a short joke"})
-
-    spans = exporter.get_finished_spans()
-
-    assert [
-        "ChatPromptTemplate.task",
-        "ChatAnthropic.chat",
-        "RunnableSequence.workflow",
-    ] == [span.name for span in spans]
-
-    anthropic_span = next(span for span in spans if span.name == "ChatAnthropic.chat")
-    workflow_span = next(
-        span for span in spans if span.name == "RunnableSequence.workflow"
-    )
-
-    assert anthropic_span.attributes[SpanAttributes.LLM_REQUEST_TYPE] == "chat"
-    assert anthropic_span.attributes[SpanAttributes.LLM_REQUEST_MODEL] == "claude-2.1"
-    assert (
-        anthropic_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.content"]
-    ) == "You are a helpful assistant"
-    assert (
-        anthropic_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.role"]
-    ) == "system"
-    assert (
-        anthropic_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.1.content"]
-    ) == "tell me a short joke"
-    assert (anthropic_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.1.role"]) == "user"
-    assert (
-        anthropic_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"]
-        == response.content
-    )
-    assert (
-        anthropic_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.role"]
-    ) == "assistant"
-    outputs = json.loads(
-        workflow_span.attributes[SpanAttributes.TRACELOOP_ENTITY_OUTPUT]
-    )["outputs"]
-    assert (
-        dict(re.findall(r'(\S+)=(".*?"|\S+)', outputs))["content"]
-        == f'"{response.content}"'
-    )
-
-
-@pytest.mark.vcr
-def test_bedrock(exporter):
-    prompt = ChatPromptTemplate.from_messages(
-        [("system", "You are a helpful assistant"), ("user", "{input}")]
-    )
-    model = BedrockChat(
-        model_id="anthropic.claude-3-haiku-20240307-v1:0",
-        client=boto3.client(
-            "bedrock-runtime",
-            aws_access_key_id="test_key",
-            aws_secret_access_key="test_secret",
-            aws_session_token="a/mock/token",
-            region_name="us-east-1",
-        ),
-    )
-
-    chain = prompt | model
-    response = chain.invoke({"input": "tell me a short joke"})
-
-    spans = exporter.get_finished_spans()
-
-    assert [
-        "ChatPromptTemplate.task",
-        "BedrockChat.chat",
-        "RunnableSequence.workflow",
-    ] == [span.name for span in spans]
-
-    bedrock_span = next(span for span in spans if span.name == "BedrockChat.chat")
-    workflow_span = next(
-        span for span in spans if span.name == "RunnableSequence.workflow"
-    )
-
-    assert bedrock_span.attributes[SpanAttributes.LLM_REQUEST_TYPE] == "chat"
-    assert (
-        bedrock_span.attributes[SpanAttributes.LLM_REQUEST_MODEL]
-        == "anthropic.claude-3-haiku-20240307-v1:0"
-    )
-    assert (
-        bedrock_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.content"]
-    ) == "You are a helpful assistant"
-    assert (bedrock_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.0.role"]) == "system"
-    assert (
-        bedrock_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.1.content"]
-    ) == "tell me a short joke"
-    assert (bedrock_span.attributes[f"{SpanAttributes.LLM_PROMPTS}.1.role"]) == "user"
-    assert (
-        bedrock_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.content"]
-        == response.content
-    )
-    assert (
-        bedrock_span.attributes[f"{SpanAttributes.LLM_COMPLETIONS}.0.role"]
-    ) == "assistant"
-    outputs = json.loads(
-        workflow_span.attributes[SpanAttributes.TRACELOOP_ENTITY_OUTPUT]
-    )["outputs"]
-    assert (
-        dict(re.findall(r'(\S+)=(".*?"|\S+)', outputs))["content"].replace("\\n", "\n")
-        == f'"{response.content}"'
-    )
